fixed error

+++ helper.py
@@ -283,6 +283,10 @@ class ResampleSpectrogram(object):
         self.output_time = torch.linspace(output_start_time, output_end_time, self.output_number_rows )  # create array that corresponds to the output time axis
         self.output_freq = torch.linspace(output_start_frequency, output_stop_frequency, self.output_number_cols)    # create array that corresponds tot the output wavelength axis
 
+        # ensure all tensors have the same type (float32)
+        self.output_time = self.output_time.float()
+        self.output_freq = self.output_freq.float()
+
     def __call__(self, spectrogram_data):
         '''
         Takes path of spectrogram and resamples it to the configured size and range of time/wavelength 
@@ -306,7 +310,11 @@ class ResampleSpectrogram(object):
         else:   # if unvalid type
             logger.error(f"type='{self.type}' is not a valid type. Use 'wavelength' or 'frequency' instead!")
             return
-
+        # ensure all tensors are of the same type (float32)
+        spectrogram = spectrogram.float()
+        input_time = input_time.float()
+        input_freq = input_freq.float()
+        
          # get minimum and maximum values of the input_time and input_freq tensors
         input_time_min, input_time_max = input_time.min(), input_time.max()
         input_freq_min, input_freq_max = input_freq.min(), input_freq.max()
